Fixed the camera issues with the View.
Added "const sf::View& view" to render() function in TextureSelector class.

Added "this->window->setView(this->view);" to renderSelectorRect() function in Editor class due to the view.

Changed view.setSize & view.setCenter to grab window size instead of graphicsSettings.resolution to stay consistent.

Changed textureSelector->update to grab mousePositionWindow instead of mousePositionView due to the view in Editor class.

Changed pause->update to grab mousePositionWindow instead of mousePositionView due to the view in Editor class.

Changed renderTiles() to renderTileMap() to be more descriptive.

Added "this->window->setView(this->window->getDefaultView());" to the render() function in Editor class for items that need to be rendered with that view.