Added bool resized; flag to State.h
Initialized this->resized = false in State constructor in State.cpp.

Removed this->initalizeMainMenu() from updateGraphicsSettings() in State.cpp.

Added this->initalizeMainMenu() to createWindow() in State.cpp.

Added this->initView(); and this->resized = true to initializeMainMenu() in State.cpp.

Added void reinitializeEditor() function to Editor.h.

Updated the Exit button if statement to the following in void updatePauseMenuButtons() in Editor.cpp:

if (this->pauseMenu->isButtonPressed("EXIT") && this->getKeyTime())
 {
  this->tileMap->saveToFile("Config/tile_map.ini");
  this->endState();
 }

Added the following to update(const float& dt) in Editor.cpp:

 if (this->resized)
  this->reinitializeEditor();

Added the following void reinitializeEditor() function to Editor.cpp:

/*Reinitialize Functions*/
void Editor::reinitializeEditor()
{
 std::cout << "Reinitializing Editor!\n";
 this->tileMap->saveToFile("Config/tile_map.ini");
 this->initVariables();
 this->initKeybinds();
 this->initFonts();
 this->initTileMap();
 this->initTextureSelector();
 this->initPauseMenu();
 this->initLatestTileMap();
 this->resized = false;
}