Updated Texture Sheet with new Tile Size of 40.f.
Initialized this->tileSize to 40.f in the State constructor in State.cpp.

Changed bound size of Texture Seletor to 1520.f, 720.f in void initTextureSelector() in Editor.ccp.

Changed this->sideBar.setSize() to the following in  void initTextureSelector() in Editor.ccp to accomidate for new Tile Size:

this->sideBar.setSize(sf::Vector2f(40.f, static_cast<float>(this->window->getSize().y)));

Changed the following in TextureSelector Constructor in TimeMap.cpp to accomidate for the new Tile Size:

/*X Position Offset*/
 float xOffset = 40.f;
 float hideButtonOffset = 20.f;

/*Hide Buttons*/
 this->hideButton = std::make_unique<GUI::Button>(
  pos_x + hideButtonOffset, pos_y + hideButtonOffset,                              //Button Rect Position
  40.f, 40.f,                                                                      //Button Rect Size
  &hide_button_font, "TS", 16,                                                     //Button Font, Text, and Character Size
  sf::Color::White, sf::Color::White, sf::Color::White,                            //Text Color
  sf::Color(70, 70, 70, 0), sf::Color(150, 150, 150, 0), sf::Color(20, 20, 20, 0), //Button Rect Fill Color
  sf::Color::White, sf::Color::White, sf::Color::White                             //Button Rect Outline Color
  );